+++ app/templates/directory.html
@@ -1,0 +1,50 @@
+{% extends 'bootstrap/base.html' %}
+
+{% block styles %}
+{{ super() }}
+	<style>
+		body { background: #e8f1f9; }
+	</style>
+{% endblock %}
+
+
+{% block title %}
+BitCan Well Directory
+{% endblock %}
+
+
+{% block content %}
+
+<div class="container">
+  <div class="row pb-5">
+    <div class="col-md-10 col-lg-8 mx-lg-auto mx-md-auto">
+
+      <h1 class="pt-5 pb-2">BitCan Well Directory</h1>
+
+      <p class="lead alert alert-warning">This is a list of all the well sites in BitCan's database
+
+	  <p class="ml-4"><a href="{{ url_for('database') }}">Return to index.</a></p>
+
+		<table class="table">
+			<tr>
+				<th>Well Name</th>
+				<th>Latitude</th>
+				<th>Longitude</th>
+			</tr>
+
+			<!-- here is a for-loop to give us all the socks -->
+			{% for well in uwi %}
+			<tr>
+				<td>{{ well.Directory }}</td>
+				<td>{{ well.Latitude }}</td>
+				<td>{{ well.Longitude }}</td>
+
+			</tr>
+			{% endfor %}
+		</table>
+
+		</div>
+	</div>
+</div>
+
+{% endblock %}
