+++ app/templates/edit_or_delete.html
@@ -1,0 +1,141 @@
+{% extends 'bootstrap/base.html' %}
+{% import "bootstrap/wtf.html" as wtf %}
+
+
+{% block styles %}
+{{ super() }}
+	<style>
+		body { background: #e8f1f9; }
+	</style>
+{% endblock %}
+
+
+{% block title %}
+	{% if choice == 'delete' %}
+		Delete a Record
+	{% elif choice == 'edit' %}
+		Edit a Sock
+	{% endif %}
+{% endblock %}
+
+
+{% block content %}
+
+<div class="container">
+  <div class="row pb-5">
+    <div class="col-lg-10 mx-lg-auto">
+
+      <h1 class="pt-5 pb-2">
+		{% if choice == 'delete' %}
+	  		Delete a Record
+	  	{% elif choice == 'edit' %}
+	  		Edit a Record
+	  	{% endif %}
+	  </h1>
+	  <!-- ----- THIS TEMPLATE HAS TWO FORMS ----- -->
+	  <!-- but only one will be visible, depending on how the template
+  		   was opened -->
+
+	  <!-- ----- start delete sock content ----- -->
+
+	  {% if choice == 'delete' %}
+
+	      <p class="lead alert alert-danger">Are you sure you want to delete this sock?</p>
+
+		  <!-- links -->
+		  <p class="ml-4">No. <a href="{{ url_for('home') }}">Return to the index.</a></p>
+
+		  <p class="ml-4">No. <a href="{{ url_for( 'select_record', letters='AF' ) }}">Return to the selection page.</a></p>
+
+		<!-- table showing the selected sock -->
+		<table class="table">
+			<tr>
+				<th>Sock Name</th>
+				<th>Color</th>
+				<th>Style</th>
+				<th>Price</th>
+				<th>Quantity</th>
+				<th>Updated</th>
+			</tr>
+
+			<!-- just one sock here -->
+			<tr>
+				<td>{{ well.Directory }}</td>
+				<td>{{ well.Shmin }}</td>
+				<td>{{ well.Type }}</td>
+				<td>{{ well.Temperature }}</td>
+				<td>{{ well.PorePressure }}</td>
+				<td>{{ well.rowid }}</td>
+			</tr>
+
+		</table>
+
+
+		<!-- start form for deleting record - not quickform but still WTForms -->
+  		<form action="{{ url_for('delete_result') }}" method="POST" id="delete">
+			{{ form2.csrf_token }}
+			{{ form2.id_field(value=sock.id) }}
+			{{ form2.purpose(value='delete') }}
+			{{ form2.submit(class="btn btn-danger space-below-btn") }}
+		</form>
+		<!-- end form -->
+
+
+		<!-- ----- end delete sock content ----- -->
+		{% else %}
+
+		<!-- ----- start edit sock content ----- -->
+
+		<p class="lead alert alert-success">Edit the selected sock below.</p>
+
+		<!-- links -->
+		<p><a href="{{ url_for('home') }}">Return to the index.</a></p>
+
+		<p><a href="{{ url_for( 'select_record', letters='AF' ) }}">Return to the selection page.</a></p>
+
+		<!-- show flash - based on WTForms validators
+			 see https://pythonprogramming.net/flash-flask-tutorial/
+			 get_flashed_messages() exists here because of flash()
+			 in the route function
+			 -->
+			 {% with errors = get_flashed_messages() %}
+				 {% if errors %}
+					 {% for err in errors %}
+					 <div class="alert alert-danger alert-dismissible" role="alert">
+						 <button type="button" class="close" data-dismiss="alert" aria-label="Close"><span aria-hidden="true">×</span></button>
+						 {{ err }}
+					 </div>
+					 {% endfor %}
+				 {% endif %}
+			 {% endwith %}
+		<!-- end of flash -->
+
+		<!-- start form for editing record
+			 correct use of WTF Bootstrap styles -->
+		<form method="post" action="{{ url_for('edit_result') }}" id="update">
+			{{ form1.csrf_token }}
+			{{ form1.id_field(value=sock.id) }}
+			{{ wtf.form_field(form1.name, value=sock.name) }}
+			{{ wtf.form_field(form1.style) }}
+			{{ wtf.form_field(form1.color, value=sock.color) }}
+			{{ wtf.form_field(form1.quantity, value=sock.quantity) }}
+			{{ wtf.form_field(form1.price, value=sock.price) }}
+			{{ wtf.form_field(form1.submit, button_map={'submit':'success'}) }}
+		</form>
+
+		<script>
+		// this JavaScript fixes the style choice in the select element
+		// without this, the current value is not shown in the menu
+		if ( document.querySelector('select') ) {
+			document.querySelector('select').value = "{{ sock.style }}";
+		}
+		</script>
+
+		{% endif %}
+
+		</div>
+	</div>
+</div>
+
+
+{% endblock %}
