+++ app/views.py
@@ -19,3 +19,7 @@ def form():
 def form2():
     return render_template("uwiForm2.html")
 
+@app.route('/database/')
+def database():
+    return render_template('database.html')
+
